pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "slideagent"
version = "0.1.0"
description = "AI agent to convert course slides + transcript into full lecture notes"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.50.0",
  "python-dotenv>=1.0.1",
  "pypdf>=5.0.0",
  "python-pptx>=1.0.2",
  "pymupdf>=1.24.0",
  "pytesseract>=0.3.10",
  "Pillow>=10.0.0",
  "pydantic>=2.9.0",
  "rich>=13.8.1",
  "click>=8.1.7",
  "streamlit>=1.39.0",
  "pytest>=8.3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
slideagent = "lecture_note_agent.cli:main"
slideagent-ui = "lecture_note_agent.ui:run"
